Today (3/11/25): I started my - Job Application Tracker project.
My idea is as follows - making a personal project, from a tech stack that lots of companies use, to then having this on your cv - to get more interviews and possibly a job!?

So, the idea of this project, is to simply make a project to keep track of the jobs that you have applied to!

I am very new to React. I am taking this project on to learn more about React and become a better frontend developer.

So, at the moment the project has the following:

A home page

A profile page - you make profile and its shown to you (no data is stored or taken, this is only a frontend project!) - I will probably remove this feature....

A applications page - this is the main core page, you enter in data that you want to track and once you have supplied all of the data, its shown to you in a nice table format.