Today (5/11/25), I have now implemented light mode & dark mode.
I have refactored my Mode.jsx file a bit, and I have also refactored my Welcome.jsx file.

I have made a new file called "style.css" this is used to set the color of the background and the color of the text depending on if its in light mode or dark mode.